Refactor new table methods
