+++ resources/js/components/shared/FilterTableBodyMixin.js
@@ -1,0 +1,29 @@
+export default {
+  methods: {
+    containsHTML(text) {
+      const doc = new DOMParser().parseFromString(text, 'text/html');
+      return Array.from(doc.body.childNodes).some(node => node.nodeType === Node.ELEMENT_NODE);
+    },
+    isComponent(content) {
+      if (content && typeof content === 'object') {
+        return content.component && typeof content.props === 'object';
+      }
+      return false;
+    },
+    sanitize(html) {
+      let cleanHtml = html.replace(/<script(.*?)>[\s\S]*?<\/script>/gi, "");
+      cleanHtml = cleanHtml.replace(/<style(.*?)>[\s\S]*?<\/style>/gi, "");
+      cleanHtml = cleanHtml.replace(
+        /<(?!b|\/b|br|img|a|input|hr|link|meta|time|button|select|textarea|datalist|progress|meter|span)[^>]*>/gi,
+        "",
+      );
+      cleanHtml = cleanHtml.replace(/\s+/g, " ");
+
+      return cleanHtml;
+    },
+    changePage(page) {
+      this.page = page;
+      this.fetch();
+    },
+  },
+};
