+++ resources/js/processes/categories/components/CategoriesListing.vue
@@ -88,12 +88,13 @@
   import dataLoadingMixin from "../../../components/common/mixins/apiDataLoading";
   import { createUniqIdsMixin } from "vue-uniq-ids";
   import EllipsisMenu from "../../../components/shared/EllipsisMenu.vue";
-  import paginationTable from "../../../components/shared/PaginationTable.vue";    
+  import paginationTable from "../../../components/shared/PaginationTable.vue";
+  import FilterTableBodyMixin from "../../../components/shared/FilterTableBodyMixin";
   const uniqIdsMixin = createUniqIdsMixin();
 
   export default {
     components: {EllipsisMenu, paginationTable},
-    mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin],
+    mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin, FilterTableBodyMixin],
     props: ["filter", "permissions", "apiRoute", "include", "labelCount", "count", "loadOnStart"],
     data () {
       return {
@@ -261,31 +262,6 @@
         };
         return bubbleColor[status];
       },
-      containsHTML(text) {
-          const doc = new DOMParser().parseFromString(text, 'text/html');
-          return Array.from(doc.body.childNodes).some(node => node.nodeType === Node.ELEMENT_NODE);
-        },
-        isComponent(content) {
-          if (content && typeof content === 'object') {
-            return content.component && typeof content.props === 'object';
-          }
-          return false;
-        },
-        sanitize(html) {
-          let cleanHtml = html.replace(/<script(.*?)>[\s\S]*?<\/script>/gi, "");
-          cleanHtml = cleanHtml.replace(/<style(.*?)>[\s\S]*?<\/style>/gi, "");
-          cleanHtml = cleanHtml.replace(
-            /<(?!b|\/b|br|img|a|input|hr|i|link|meta|time|button|select|textarea|datalist|progress|meter|span)[^>]*>/gi,
-            "",
-          );
-          cleanHtml = cleanHtml.replace(/\s+/g, " ");
-
-          return cleanHtml;
-        },
-        changePage(page) {
-          this.page = page;
-          this.fetch();
-        },
     }
   };
 </script>
