+++ resources/js/templates/components/ProcessTemplatesListing.vue
@@ -94,13 +94,14 @@
     import { createUniqIdsMixin } from "vue-uniq-ids";
     import EllipsisMenu from "../../components/shared/EllipsisMenu.vue";
     import { FilterTable } from "../../components/shared";
-    import paginationTable from "../../components/shared/PaginationTable.vue";    
+    import paginationTable from "../../components/shared/PaginationTable.vue";
+    import FilterTableBodyMixin from "../../components/shared/FilterTableBodyMixin";
   
     const uniqIdsMixin = createUniqIdsMixin();
   
     export default {
-      components: {EllipsisMenu, paginationTable},
-      mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin],
+      components: { EllipsisMenu, paginationTable },
+      mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin, FilterTableBodyMixin],
       props: ["filter", "id", "status", "permission", "isDocumenterInstalled", "processName"],
       data() {
         return {
@@ -315,31 +316,6 @@
           });
           return data;
         },
-        containsHTML(text) {
-          const doc = new DOMParser().parseFromString(text, 'text/html');
-          return Array.from(doc.body.childNodes).some(node => node.nodeType === Node.ELEMENT_NODE);
-        },
-        isComponent(content) {
-          if (content && typeof content === 'object') {
-            return content.component && typeof content.props === 'object';
-          }
-          return false;
-        },
-        sanitize(html) {
-          let cleanHtml = html.replace(/<script(.*?)>[\s\S]*?<\/script>/gi, "");
-          cleanHtml = cleanHtml.replace(/<style(.*?)>[\s\S]*?<\/style>/gi, "");
-          cleanHtml = cleanHtml.replace(
-            /<(?!b|\/b|br|img|a|input|hr|link|meta|time|button|select|textarea|datalist|progress|meter|span)[^>]*>/gi,
-            "",
-          );
-          cleanHtml = cleanHtml.replace(/\s+/g, " ");
-
-          return cleanHtml;
-        },
-        changePage(page) {
-          this.page = page;
-          this.fetch();
-        },        
       },
     };
   </script>
